+++ src/utils/useHasSubgraphSyncErrors.ts
